requirements.txt
# Fusion Standalone - RL Pentest Training Framework
# Merged dependencies from Script and PenGym

# ============ Core ML ============
torch>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ============ NLP & Embeddings ============
sentence-transformers>=2.2.0
transformers>=4.36.0
tokenizers>=0.15.0

# ============ RL Environment ============
nasim>=0.10.0
gymnasium>=0.29.0

# ============ PenGym Tools (for real execution) ============
python-nmap>=0.7.0
pymetasploit3>=1.0.0
psutil>=5.9.0

# ============ Visualization & Logging ============
tensorboard>=2.15.0
wandb>=0.16.0
matplotlib>=3.8.0
tqdm>=4.66.0

# ============ Utilities ============
pyyaml>=6.0.0
easydict>=1.11
loguru>=0.7.0
colorama>=0.4.6
rich>=13.0.0
pandas>=2.1.0
prettytable>=3.9.0
